Users can signup, authenticate and get profile

+++ database/migrations/1503250034280_token.js
@@ -4,10 +4,10 @@
 const Schema = use('Schema')
 
 class TokensSchema extends Schema {
-  up () {
+  up() {
     this.create('tokens', (table) => {
       table.increments()
-      table.integer('user_id').unsigned().references('id').inTable('users')
+      table.bigInteger('user_id').unsigned().references('id').inTable('users')
       table.string('token', 255).notNullable().unique().index()
       table.string('type', 80).notNullable()
       table.boolean('is_revoked').defaultTo(false)
@@ -15,7 +15,7 @@ class TokensSchema extends Schema {
     })
   }
 
-  down () {
+  down() {
     this.drop('tokens')
   }
 }
